Add files via upload
I tested the uploaded file on a computer with older R packages. The versions tested are listed in the R script. The problem seems to be caused by the packages loaded at the the start of R which cannot be restored to an earlier version as the other packages in use were replaced by the versions on the computer with the older R package.
The error message caused with ggpredict is:
pred2 <- ggpredict(mod2,terms="JAHR")
Error in `vectbl_as_col_location2()`:
! Can't extract column with `terms[2]`.
✖ Subscript `terms[2]` must be a location, not a character `NA`.
Backtrace:
  1. ggeffects::ggpredict(mod2, terms = "JAHR")
  2. ggeffects:::ggpredict_helper(...)
  3. ggeffects:::.post_processing_predictions(...)
  4. ggeffects:::.add_labels_to_groupvariable(...)
  7. tibble:::`[[.tbl_df`(original_model_frame, terms[2])
  8. tibble:::tbl_subset2(x, j = i, j_arg = substitute(i))
  9. tibble:::vectbl_as_col_location2(j, length(x), names(x), j_arg = j_arg)
 12. vctrs::vec_as_location2(j, n, names)
 13. vctrs:::result_get(...)